+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/flymodes/aac/AAC3312Glide.kt
@@ -7,7 +7,7 @@ package net.ccbluex.liquidbounce.features.module.modules.movement.flymodes.aac
 
 import net.ccbluex.liquidbounce.features.module.modules.movement.flymodes.FlyMode
 
-object AAC3312Glide : FlyMode("AAG3.3.12-Glide") {
+object AAC3312Glide : FlyMode("AAC3.3.12-Glide") {
     private var tick = 0
 
     override fun onUpdate() {
@@ -15,12 +15,12 @@ object AAC3312Glide : FlyMode("AAG3.3.12-Glide") {
             if (!onGround)
                 tick++
 
-            when (tick) {
-                2 -> mc.timer.timerSpeed = 1f
-                in 12..13 && !onGround -> {
+            when {
+                tick == 2 -> mc.timer.timerSpeed = 1f
+                tick >= 12 && !onGround -> {
                     mc.timer.timerSpeed = 0.1f
-                    tick = 0
                     motionY = 0.015
+                    tick = 0
                 }
             }
         }